# contracts/openapi/v1.yaml
openapi: 3.1.0
info:
  title: 命理工具 API
  version: 0.1.0
  description: Identity, profile, reading and presentation contract.
servers:
  - url: /
    description: Same-origin web entry
...
paths:
...
  /api/v1/account:
    get:
      operationId: getAccount
      tags: [Identity]
      summary: Read the current internal User and verified Login Identities
      security:
        - deviceSession: []
      responses:
        "200":
          description: Current account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/account/export:
    get:
      operationId: exportAccountData
      tags: [Identity]
      summary: Export the current account data and decrypted profile payloads
      security:
        - deviceSession: []
      responses:
        "200":
          description: Account data export
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataExportResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v1/account/notifications:
    get:
      operationId: listAccountNotifications
      tags: [Identity]
      summary: List private in-app notifications for the current account
      security:
        - deviceSession: []
      parameters:
        - name: unread_only
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: Safe in-app notification projections without delivery payloads
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountNotificationsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/account/notifications/read-all:
    post:
      operationId: markAllAccountNotificationsRead
      tags: [Identity]
      summary: Mark every visible in-app notification as read
      security:
        - deviceSession: []
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "200":
          description: Remaining unread count after the operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountNotificationsReadAllResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/account/notifications/{notification_id}/read:
    post:
      operationId: markAccountNotificationRead
      tags: [Identity]
      summary: Mark one owned in-app notification as read
      security:
        - deviceSession: []
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - name: notification_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Updated safe in-app notification projection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountNotificationResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/account/notifications/{notification_id}:
    delete:
      operationId: deleteAccountNotification
      tags: [Identity]
      summary: Soft-delete one owned in-app notification
      security:
        - deviceSession: []
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - name: notification_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Notification removed from the current account view
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/account/notification-preferences:
    get:
      operationId: getNotificationPreferences
      tags: [Identity]
      summary: Read the current account's notification channel preferences
      security:
        - deviceSession: []
      responses:
        "200":
          description: Notification preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferencesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      operationId: updateNotificationPreferences
      tags: [Identity]
      summary: Update the current account's notification channel preferences
      security:
        - deviceSession: []
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationPreferencesRequest"
      responses:
        "200":
          description: Updated notification preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferencesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/account/closure:
    get:
      operationId: getAccountClosure
      tags: [Identity]
      summary: Read the active account closure request
      security:
        - deviceSession: []
      responses:
        "200":
          description: Active request or null
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: "#/components/schemas/ClosureResponse"
                  - type: "null"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: requestAccountClosure
      tags: [Identity]
      summary: Request account closure with a seven-day cancellation window
      security:
        - deviceSession: []
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "201":
          description: Closure requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClosureResponse"
        "200":
          description: Existing pending request returned idempotently
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClosureResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      operationId: cancelAccountClosure
      tags: [Identity]
      summary: Cancel account closure before the seven-day deadline
      security:
        - deviceSession: []
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "204":
          description: Closure cancelled
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    deviceSession:
      type: apiKey
      in: cookie
      name: mingli_session
  parameters:
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 32
...
  schemas:
...
    ClosureResponse:
      type: object
      additionalProperties: false
      required: [closure_id, user_id, status, requested_at, cancel_until]
      properties:
        closure_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, cancelled, executed]
        requested_at:
          type: string
          format: date-time
        cancel_until:
          type: string
          format: date-time
        cancelled_at:
          type: [string, "null"]
          format: date-time
        executed_at:
          type: [string, "null"]
          format: date-time
    DataExportResponse:
      type: object
      additionalProperties: false
      required: [generated_at, user_id, payload]
      properties:
        generated_at:
          type: string
          format: date-time
        user_id:
          type: string
          format: uuid
        payload:
          type: object
          additionalProperties: true
...
    LoginIdentitySummary:
      type: object
      additionalProperties: false
      required: [id, provider, masked_destination, verified_at]
      properties:
        id:
          type: string
          format: uuid
        provider:
          type: string
          enum: [phone, email]
        masked_destination:
          type: string
        verified_at:
          type: string
          format: date-time
    AccountResponse:
      type: object
      additionalProperties: false
      required: [user_id, identities]
      properties:
        user_id:
          type: string
          format: uuid
        identities:
          type: array
          items:
            $ref: "#/components/schemas/LoginIdentitySummary"
    NotificationPreferencesRequest:
      type: object
      additionalProperties: false
      required: [in_app_enabled, email_enabled, sms_enabled]
      properties:
        in_app_enabled:
          type: boolean
        email_enabled:
          type: boolean
        sms_enabled:
          type: boolean
    NotificationPreferencesResponse:
      type: object
      additionalProperties: false
      required: [in_app_enabled, email_enabled, sms_enabled]
      properties:
        in_app_enabled:
          type: boolean
        email_enabled:
          type: boolean
        sms_enabled:
          type: boolean
    AccountNotification:
      type: object
      additionalProperties: false
      required: [id, title, summary, available_at, read_at, target_href]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        summary:
          type: string
        available_at:
          type: string
          format: date-time
        read_at:
          type: [string, "null"]
          format: date-time
        target_href:
          type: [string, "null"]
    AccountNotificationsResponse:
      type: object
      additionalProperties: false
      required: [notifications, unread_count]
      properties:
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/AccountNotification"
        unread_count:
          type: integer
          minimum: 0
    AccountNotificationsReadAllResponse:
      type: object
      additionalProperties: false
      required: [unread_count]
      properties:
        unread_count:
          type: integer
          minimum: 0
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, request_id]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        request_id:
          type: string
  responses:
...
    Unauthorized:
      description: The operation's required Device or Guest owner session is missing or invalid; an invalid submitted Device Session expires stale device and CSRF cookies
      headers:
        Set-Cookie:
          description: May expire stale mingli_session and mingli_csrf cookies before re-authentication
          schema:
            type: string
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: CSRF or origin validation failed
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: The resource does not exist or belongs to another owner
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Conflict:
      description: The resource is in a state that does not allow this transition
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...